examples/rabbitmq/example-app/multi-agent-app/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "multi-agent-app"
version = "0.1.0"
description = ""
authors = ["Andrei Fajardo <andrei@runllama.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
llama-agents = {version = "^0.0.5", extras = ["rabbitmq"]}
llama-index-llms-openai = "^0.1.25"
llama-index-embeddings-openai = "^0.1.10"
llama-index-agent-openai = "^0.2.7"
